Split controller.go off of the HPA reconciler
